move prompts to a single file
